+++ packages/vtable-plugins/demo/menu.ts
@@ -40,6 +40,10 @@ export const menus = [
     name: 'add-row-column'
   },
   {
+    path: 'paste-add-row-column',
+    name: 'paste-add-row-column'
+  },
+  {
     path: 'column-series',
     name: 'column-series'
   },
@@ -52,6 +56,38 @@ export const menus = [
     name: 'rotate-table'
   },
   {
+    path: 'master-detail-plugin',
+    name: 'master-detail-plugin'
+  },
+  {
+    path: 'master-detail-plugin',
+    name: 'master-detail-plugin2'
+  },
+  {
+    path: 'master-detail-plugin',
+    name: 'master-detail-plugin3'
+  },
+  {
+    path: 'master-detail-plugin',
+    name: 'master-detail-plugin4'
+  },
+  {
+    path: 'master-detail-plugin',
+    name: 'master-detail-plugin5'
+  },
+  {
+    path: 'master-detail-plugin',
+    name: 'master-detail-plugin6'
+  },
+  {
+    path: 'master-detail-plugin',
+    name: 'master-detail-plugin7'
+  },
+  {
+    path: 'master-detail-plugin',
+    name: 'master-detail-plugin8'
+  },
+  {
     menu: 'pivot-plugin',
     children: [
       {
@@ -59,5 +95,9 @@ export const menus = [
         name: 'pivot-plugin'
       }
     ]
+  },
+  {
+    path: 'table-export',
+    name: 'table-export'
   }
 ];