+++ packages/vtable-plugins/src/master-detail-plugin.ts
@@ -1,0 +1,5 @@
+export {
+  MasterDetailPlugin,
+  type MasterDetailPluginOptions,
+  type DetailGridOptions
+} from './master-detail-plugin/index';