+++ packages/vtable-plugins/src/table-export/util/promise.ts
@@ -1,0 +1,3 @@
+export function isPromise(obj: any) {
+  return !!obj && (typeof obj === 'object' || typeof obj === 'function') && typeof obj.then === 'function';
+}