+++ packages/vtable/src/ts-types/events.ts
@@ -12,8 +12,8 @@ import type { DropDownMenuEventArgs, MenuListItem, PivotInfo } from './menu';
 import type { IDimensionInfo, MergeCellInfo, RectProps, SortOrder } from './common';
 import type { IconFuncTypeEnum, CellInfo, HierarchyState } from '.';
 import type { Icon } from '../scenegraph/graphic/icon';
-import type { FederatedPointerEvent, IEventTarget } from '@src/vrender';
-import type { BaseTableConstructorOptions } from './base-table';
+import type { FederatedPointerEvent, Group, IEventTarget } from '@src/vrender';
+import type { BaseTableConstructorOptions, BaseTableAPI } from './base-table';
 
 export type KeyboardEventListener = (e: KeyboardEvent) => void;
 export type TableEventListener<TYPE extends keyof TableEventHandlersEventArgumentMap> = (
@@ -241,6 +241,34 @@ export interface TableEventHandlersEventArgumentMap {
   before_set_size: { width: number; height: number };
   after_render: null;
   initialized: null;
+  after_update_cell_content_width: {
+    col: number;
+    row: number;
+    distWidth: number;
+    cellHeight: number;
+    detaX: number;
+    autoRowHeight: boolean;
+    needUpdateRowHeight: boolean;
+    cellGroup: Group;
+    padding: [number, number, number, number];
+    textAlign: CanvasTextAlign;
+    textBaseline: CanvasTextBaseline;
+  };
+  before_create_progress_bar: {
+    col: number;
+    row: number;
+    width: number;
+    height: number;
+    table: BaseTableAPI;
+    range?: CellRange;
+    modifiedHeight: number;
+  };
+  after_update_select_border_height: {
+    startRow: number;
+    endRow: number;
+    currentHeight: number;
+    selectComp: { rect: any; fillhandle?: any; role: string };
+  };
 
   change_cell_value: {
     col: number;
@@ -350,6 +378,8 @@ export interface TableEventHandlersReturnMap {
   before_set_size: void;
   after_render: void;
   initialized: void;
+  after_update_cell_content_width: void;
+  before_create_progress_bar: void;
 
   change_cell_value: void;
   mousedown_fill_handle: void;