+++ packages/vtable/src/ts-types/table-engine.ts
@@ -3,7 +3,13 @@ import type { ColumnIconOption, SvgIcon } from './icon';
 export type { HeaderData } from './list-table/layout-map/api';
 export type LayoutObjectId = number | string;
 import type { Rect } from '../tools/Rect';
-import type { BaseTableAPI, BaseTableConstructorOptions, ListTableProtected, PivotTableProtected } from './base-table';
+import type {
+  BaseTableAPI,
+  BaseTableConstructorOptions,
+  ListTableProtected,
+  PivotChartProtected,
+  PivotTableProtected
+} from './base-table';
 import type {
   Aggregation,
   AggregationType,
@@ -138,8 +144,9 @@ export interface IRowSeriesNumber {
   // align?: 'left' | 'right';
   // span?: number | 'dependOnNear';
   title?: string;
-  field?: FieldDef;
+  field?: string | number;
   format?: (col?: number, row?: number, table?: BaseTableAPI) => any;
+  headerType?: 'text' | 'link' | 'image' | 'video' | 'checkbox';
   cellType?: 'text' | 'link' | 'image' | 'video' | 'checkbox' | 'radio';
   style?: ITextStyleOption | ((styleArg: StylePropertyFunctionArg) => ITextStyleOption);
   headerStyle?: ITextStyleOption | ((styleArg: StylePropertyFunctionArg) => ITextStyleOption);
@@ -300,10 +307,24 @@ export interface ListTableConstructorOptions extends BaseTableConstructorOptions
       }) => Aggregation | CustomAggregation | (Aggregation | CustomAggregation)[] | null);
   /** 数据为空时显示聚合结果 */
   showAggregationWhenEmpty?: boolean;
+  /** 针对column中配置了tree: true的列，开启这个配置后，可以合并分组标题。需要配合在数据中配置vtableMerge和vtableMergeName。默认为false */
   enableTreeNodeMerge?: boolean;
+  groupConfig?: {
+    groupBy: GroupByOption;
+    titleCustomLayout?: ICustomLayout;
+    titleFieldFormat?: (record: any, col?: number, row?: number, table?: BaseTableAPI) => string;
+    /** 开启分组标题吸附功能。 */
+    enableTreeStickCell?: boolean;
+    /** 这个配置对应当在rowSeriesNumber中配置cellType: 'checkbox'时，如想在group分组名中显示checkbox，则需要开启这个配置 。默认为false*/
+    titleCheckbox?: boolean;
+  };
+  /** @deprecated 请使用groupConfig */
   groupBy?: GroupByOption;
+  /** @deprecated 请使用groupConfig */
   groupTitleCustomLayout?: ICustomLayout;
+  /** @deprecated 请使用groupConfig */
   groupTitleFieldFormat?: (record: any, col?: number, row?: number, table?: BaseTableAPI) => string;
+  /** @deprecated 请使用groupConfig */
   enableTreeStickCell?: boolean;
 
   columnWidthConfig?: { key: string; width: number }[];
@@ -412,7 +433,6 @@ export interface ListTableAPI extends BaseTableAPI {
    */
   getBodyRowIndexByRecordIndex: (index: number | number[]) => number;
 
-  _parseColumnWidthConfig: (columnWidthConfig: { key: string; width: number }[]) => void;
   _hasHierarchyTreeHeader: () => boolean;
 }
 export interface MasterDetailTableAPI extends BaseTableAPI {
@@ -620,6 +640,15 @@ export interface PivotChartConstructorOptions extends BaseTableConstructorOption
     columnResizeType?: 'column' | 'indicator' | 'all' | 'indicatorGroup';
     rowResizeType?: 'row' | 'indicator' | 'all' | 'indicatorGroup';
   } & BaseTableConstructorOptions['resize'];
+
+  columnWidthConfig?: {
+    dimensions: IDimensionInfo[];
+    width: number;
+  }[];
+  columnWidthConfigForRowHeader?: {
+    dimensions: IDimensionInfo[];
+    width: number;
+  }[];
 }
 export interface PivotTableAPI extends BaseTableAPI {
   internalProps: PivotTableProtected;
@@ -644,17 +673,34 @@ export interface PivotTableAPI extends BaseTableAPI {
    * @param values 多个单元格的数据数组
    */
   changeCellValues: (col: number, row: number, values: (string | number)[][], workOnEditableCell: boolean) => void;
-  _parseColumnWidthConfig: (columnWidthConfig: { dimensions: IDimensionInfo[]; width: string | number }[]) => void;
-  _parseColumnWidthConfigForRowHeader: (
-    columnWidthConfig: { dimensions: IDimensionInfo[]; width: string | number }[]
-  ) => void;
+
+  getCellAddressByHeaderPaths: (
+    dimensionPaths:
+      | {
+          colHeaderPaths: IDimensionInfo[];
+          rowHeaderPaths: IDimensionInfo[];
+          cellLocation: CellLocation;
+        }
+      | IDimensionInfo[]
+  ) => CellAddress;
 }
 export interface PivotChartAPI extends BaseTableAPI {
+  internalProps: PivotChartProtected;
   records?: any | Record<string, any[]>;
   options: PivotChartConstructorOptions;
   // internalProps: PivotTableProtected;
   isListTable: () => false;
   isPivotTable: () => true;
+
+  getCellAddressByHeaderPaths: (
+    dimensionPaths:
+      | {
+          colHeaderPaths: IDimensionInfo[];
+          rowHeaderPaths: IDimensionInfo[];
+          cellLocation: CellLocation;
+        }
+      | IDimensionInfo[]
+  ) => CellAddress;
 }
 export type SetPasteValueTestData = CellAddress & {
   table: BaseTableAPI;